+++ CHANGELOG.md
@@ -10,9 +10,20 @@ template engine project.
 ---
 
 
-## [unreleased] - 202x-xx-xx
+## [unreleased] - 202y-mm-dd
 
-*List of modifications for a future release.*
+*Changes for the upcoming new version.*
+
+
+## [1.1.0] - 2025-01-19
+
+### Added
+
+- Add option to fill the template using a data defined by a dictionary.
+- Add option to set multiple subblocks at once using their names instead of providing the subblock
+  objects.
+- Add option to clear a single block using by the `set()` method with a `variation_idx` set to a
+  negative or `False` value. 
 
 
 ## [1.0.0] - 2024-01-05
@@ -21,4 +32,5 @@ template engine project.
 
 
 [unreleased]: https://github.com/lubomilko/blocky
+[1.1.0]: https://github.com/lubomilko/blocky/releases/tag/1.1.0
 [1.0.0]: https://github.com/lubomilko/blocky/releases/tag/1.0.0